+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.2.0
 - Add support of `package-lock.json` version 3.
+- Amend information and error messages for better skimming.
 
 ### 0.1.1
 - Fix the error accessing undefined.